[tests] Use stdint.h in 0107-bnot.c
This test was assuming several sizes in the integer types
which is totall wrong for a test.

+++ tests/execute/0107-bnot.c
@@ -1,10 +1,11 @@
 
+#include <stdint.h>
 
 int
 main()
 {
-	int x;
-	long long l;
+	int32_t x;
+	int64_t l;
 	
 	x = 0;
 	l = 0;
